requirements.txt
python-telegram-bot==20.3
selenium==4.13.0
playwright==1.42.0